urlbuilder/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
centaurminer==0.1.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.4.0
dataclasses==0.7
google-api-core==1.22.2
google-auth==1.22.0
google-cloud-bigquery==2.0.0
google-cloud-core==1.4.1
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
idna==2.10
idna-ssl==1.1.0
libcst==0.3.11
multidict==4.7.6
mypy-extensions==0.4.3
numpy==1.19.2
pandas==1.1.2
proto-plus==1.10.0
protobuf==3.13.0
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
rsa==4.6
selenium==3.141.0
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
webdriver-manager==3.2.2
yarl==1.6.0
